// src/entities/shift/model/types.ts
export interface Shift {
  id: string;
  logo: string;
  address: string;
  companyName: string;
  dateStartByCity: string;
  timeStartByCity: string;
  timeEndByCity: string;
  currentWorkers: number;
  planWorkers: number;
  workTypes: string[];
  priceWorker: number;
  customerFeedbacksCount: number;
  customerRating: number;
}

export interface ShiftsResponse {
  data: Shift[];
}